+++ src/redux/createReducer.js
@@ -1,4 +1,0 @@
-/* eslint-disable */
-// fixme: deprecated, remove in next version
-export { default } from '../createReducer';
-console.warn('@triniti/app/redux/createReducer has moved, use @triniti/app/createReducer instead.');